rag_controller/requirements.txt
fastapi==0.110.1
llama-index==0.10.59
requests==2.32.3
uvicorn==0.29.0
weaviate-client==3.26.6
